// packages/discord/src/operations/pollExpire.ts
export const PollExpireInput = /*@__PURE__*/ /*#__PURE__*/ Schema.Struct({
  channel_id: Schema.String.pipe(T.PathParam()),
  message_id: Schema.String.pipe(T.PathParam()),
}).pipe(
  T.Http({
    method: "POST",
    path: "/channels/{channel_id}/polls/{message_id}/expire",
  }),
);
export type PollExpireInput = typeof PollExpireInput.Type;
// ...
export type PollExpireOutput = typeof PollExpireOutput.Type;

// The operation
export const pollExpire = /*@__PURE__*/ /*#__PURE__*/ API.make(() => ({
  inputSchema: PollExpireInput,
  outputSchema: PollExpireOutput,
  errors: [BadRequest, Forbidden, NotFound] as const,
}));